Removed composite ragdoll class to improve code quality
- EnemyLight subclasses caused a paralell inheritance hierarchy with ragdoll composite subclasses
- EnemyLight now holds ragdoll fields and methods, and complexity has thus been decreased
