+++ Assets/Scripts/Enemy/Enemies/Skeleton/SkeletonController.cs
@@ -23,9 +23,12 @@ public class SkeletonController : EnemyLight
     [SerializeField]
     private AudioSource audioSource = default;
 
+
     private static readonly int attackAnimatorParam = Animator.StringToHash("Attack");
     private static readonly int chasingAnimatorParam = Animator.StringToHash("Chasing");
     private static readonly int walkModeAnimatorParam = Animator.StringToHash("WalkMode");
+    private static readonly int deadAnimatorParam = Animator.StringToHash("Dead");
+    private static readonly int deathModeAnimatorParam = Animator.StringToHash("DeathMode");
 
     protected override void Start()
     {
@@ -48,6 +51,14 @@ public class SkeletonController : EnemyLight
     {
         // Set random walk animation
         anim.SetFloat(walkModeAnimatorParam, Random.Range(0, 3));
+    }
+    protected override void SetRagdollAnimatorParams()
+    {
+        base.SetRagdollAnimatorParams();
+
+        anim.SetBool(deadAnimatorParam, true);
+        // Set random death animation
+        anim.SetFloat(deathModeAnimatorParam, Random.Range(0, 8));
     }
 
     void FixedUpdate()
