--- Assets/Scripts/Enemy/Enemies/Skeleton/SkeletonRagdollController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 44aeb75c17c6bb242b9da7bec99c01aa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
