+++ Assets/Scripts/Enemy/Enemies/Unigolem/UnigolemController.cs
@@ -69,4 +69,12 @@ public class UnigolemController : EnemyLight
         if (Vector3.Distance(transform.position, baseTransform.position) < detonationDistance)
             baseHealth?.OnReceiveDamage(this, attackDamage);
     }
+
+    protected override Vector3 GetLaunchTorque()
+    {
+        // Using `forward` here would have made the most sense to me
+        // (getting the body to rotate in the same direction as this collider is facing),
+        // but that would have caused the body to rotate in its left-facing direction
+        return agent.speed * colliderToEnable.transform.right;
+    }
 }
