--- Assets/Scripts/Enemy/Enemies/Unigolem/UnigolemRagdollController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 038f81e2cfae591409b8a2244bbe762f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
