--- Assets/Scripts/Enemy/EnemyLightRagdollController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 346f67067ccd5234bad55c3ea351f17c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
